fix all mappings

+++ templates/noKibana/blocks.go
@@ -45,7 +45,7 @@ var Blocks = Object{
 					},
 					"prevEpochStartRound": Object{
 						"index": "false",
-						"type":  "long",
+						"type":  "double",
 					},
 					"rewardsForProtocolSustainability": Object{
 						"index": "false",
@@ -89,7 +89,7 @@ var Blocks = Object{
 					},
 					"nonce": Object{
 						"index": "false",
-						"type":  "long",
+						"type":  "double",
 					},
 					"pendingMiniBlockHeaders": Object{
 						"properties": Object{
@@ -122,7 +122,7 @@ var Blocks = Object{
 					},
 					"round": Object{
 						"index": "false",
-						"type":  "long",
+						"type":  "double",
 					},
 					"scheduledRootHash": Object{
 						"index": "false",
@@ -166,7 +166,7 @@ var Blocks = Object{
 				"type": "keyword",
 			},
 			"nonce": Object{
-				"type": "long",
+				"type": "double",
 			},
 			"notarizedBlocksHashes": Object{
 				"type": "keyword",
@@ -186,7 +186,7 @@ var Blocks = Object{
 				"type":  "keyword",
 			},
 			"round": Object{
-				"type": "long",
+				"type": "double",
 			},
 			"scheduledData": Object{
 				"properties": Object{
